Alter Size type.

+++ typedefs.h
@@ -11,7 +11,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-using Size = uint32_t;
+using Size = uint16_t;
 using Address = uint16_t;
 using UUID = uint64_t;
